changed folder structure resolved issues
