+++ Week-10/url-shortner/package.json
@@ -2,7 +2,6 @@
   "name": "url-shortner",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://sharunspi.github.io/p6-yaswanth-sharunspi/tree/week-10/Week-10/",
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
@@ -10,18 +9,15 @@
     "axios": "^1.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.5.0",
+    "react-router-dom": "^6.6.0",
     "react-scripts": "5.0.1",
-    "react-spinners": "^0.13.7",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
@@ -40,8 +36,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "gh-pages": "^4.0.0"
   }
 }
