+++ Week-10/url-shortner/src/Components/Footer/inex.js
